make captcha work in an svn working copy
git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@6560 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ lib/spamprotection/captcha.class.php
@@ -80,7 +80,7 @@ class Captcha {
         $handleDir = opendir($this->strAbsolutePath);
         if ($handleDir) {
             while ($strFile = readdir($handleDir)) {
-                if ($strFile != '.' && $strFile != '..' && (filemtime($this->strAbsolutePath.$strFile) < time()-3600)) {
+                if ($strFile != '.' && $strFile != '..' && $strFile != '.svn' && (filemtime($this->strAbsolutePath.$strFile) < time()-3600)) {
                     unlink($this->strAbsolutePath.$strFile);
                 }
             }
